Renamed the PlaceHolder content file to "Rankings" and moved it to the models directory.
As it stands, we will have a Rankings class and file that is responsible for holding a list of Ranking instances. I hope that's not too confusing; it'll be a pain to change once everyone starts working...
